Move get_model_config_writer to writers.py
