+++ writers/__init__.py
@@ -32,27 +32,7 @@ from writers.writers import S3ConfigWriter
 from writers.writers import GCSConfigWriter
 from writers.writers import MonitoringConfigWriter
 from writers.writers import BatchConfigWriter
-
-
-def get_model_config_writer(bucket):
-    """Based on the bucket address, return the appropriate ConfigWriter class.
-
-    Args:
-        bucket (str): Path of the storage bucket to use.
-
-    Returns:
-        ModelConfigWriter: Class to read the bucket and create a model config.
-    """
-    b = str(bucket).lower()
-    if b.startswith('s3://'):
-        return S3ConfigWriter
-
-    if b.startswith('gs://'):
-        return GCSConfigWriter
-
-    protocol = b.split('://')[0]
-    raise ValueError('Unknown bucket protocol "{}" in bucket "{}"'.format(
-        protocol, b))
+from writers.writers import get_model_config_writer
 
 
 del absolute_import
